+++ src/main/java/com/accounts/repo/BankAccountRepo.java
@@ -3,5 +3,5 @@ package com.accounts.repo;
 import com.accounts.entity.BankAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BankAccountRepo extends JpaRepository<BankAccount, Long> {
+public interface BankAccountRepo extends JpaRepository<BankAccount, Integer> {
 }
